discord-bot/requirements.txt
# Discord Bot Dependencies

discord.py>=2.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
openai>=1.0.0
